% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipa.R
\name{tipa}
\alias{tipa}
\title{Calculate the phase shift induced by a stimulus during a circadian time-course.}
\usage{
tipa(extremumTimes, switchTime, tau = NULL)
}
\arguments{
\item{extremumTimes}{Vector of times of the chosen phase reference point.}

\item{switchTime}{Number corresponding to the time of the stimulus.}

\item{tau}{Optional list with elements "pre" and "post" corresponding to the period of
the oscillator prior to and subsequent to the stimulus. If not supplied, the periods
for pre- and post-stimulus are calculated as the mean time between occurrences of the
phase reference point within the respective epoch.}
}
\value{
A list.
\item{tau}{List containing the period of the oscillator pre- and post-stimulus.}
\item{extremaPost}{Data frame of intermediate calculations.}
\item{deltaPhi}{Estimated phase shift, in units of the original measurements. Negative values
indicate a delay, positive values an advance.}
}
\description{
\code{tipa} calculates the phase shift based on the times of a phase reference point
(e.g., peak in bioluminescence) and the time of the stimulus. \code{tipa} accounts
for possible period changes and for the stage of the circadian cycle at which the stimulus
occurs, as described in \url{https://hugheylab.org/papers}.
}
\examples{
# Peak times of bioluminescence (in hours)
extremumTimes = c(-75.5, -51.5, -27.4, -3.8,
		  20.5, 42.4, 65.5, 88.0)
result = tipa(extremumTimes, switchTime=0)
result

# Convert phase shift estimate to circadian hours
result$deltaPhi / result$tau$post * 24

# Example using csv files with data from multiple experiments
extrFile = system.file('extdata', 'extremumTimes.csv', package='tipa')
switchFile = system.file('extdata', 'switchTimes.csv', package='tipa')

file.show(extrFile)
file.show(switchFile)

extrDf = read.csv(extrFile, stringsAsFactors=FALSE)
switchDf = read.csv(switchFile, stringsAsFactors=FALSE)

resultList = lapply(switchDf$expId, function(ii) {
   tipa(extrDf$extremumTime[extrDf$expId==ii],
   switchDf$switchTime[switchDf$expId==ii])
})

deltaPhiVec = sapply(resultList, function(r) r$deltaPhi)

write.csv(data.frame(expId = switchDf$expId, deltaPhi = deltaPhiVec),
	  'tipa_phase_shifts.csv', quote=FALSE, row.names=FALSE)
}
