% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipa.R
\name{tipaPhaseRef}
\alias{tipaPhaseRef}
\title{Calculate the phase shift induced by a stimulus during a circadian time-course.}
\usage{
tipaPhaseRef(phaseRefTimes, switchTime, switchDuration = 0, period = NULL)
}
\arguments{
\item{phaseRefTimes}{Vector of times of the chosen phase reference point.}

\item{switchTime}{Time at which the stimulus started.}

\item{switchDuration}{Duration of the stimulus and any transients. Data between
\code{switchTime} and \code{switchTime + switchDuration} will be ignored.}

\item{period}{Optional list with elements "pre" and "post" corresponding to the period of
the oscillations prior to and subsequent to the stimulus. If not supplied, the periods
for pre- and post-stimulus are calculated as the mean time between occurrences of the
phase reference point within the respective epoch. It is recommended to not use this
argument.}
}
\value{
A list.
\item{phaseShift}{Estimated phase shift in circadian hours. Negative values
indicate a delay, positive values an advance.}
\item{epochInfo}{Dataframe containing estimated period for each epoch.}
}
\description{
\code{tipaPhaseRef} calculates the phase shift based on the times of a phase reference point
(e.g., onset of activity), accounting for possible period changes and for the point in
the circadian cycle at which the stimulus occurred. If the rhythms of the measurement
are approximately sinusoidal, it is recommended to instead use \code{\link{tipaCosinor}}.
}
\examples{
# Peak times of bioluminescence (in hours)
phaseRefTimes = c(-75.5, -51.5, -27.4, -3.8,
		  20.5, 42.4, 65.5, 88.0)
result = tipaPhaseRef(phaseRefTimes, switchTime=0)
result

# Data from multiple (simulated) experiments
extrFile = system.file('extdata', 'phaseRefTimes.csv', package='tipa')
switchFile = system.file('extdata', 'switchTimes.csv', package='tipa')

file.show(extrFile)
file.show(switchFile)

extrDf = read.csv(extrFile, stringsAsFactors=FALSE)
switchDf = read.csv(switchFile, stringsAsFactors=FALSE)

resultList = lapply(switchDf$expId, function(ii) {
	phaseRefTimes = extrDf$phaseRefTime[extrDf$expId==ii]
	switchTime = switchDf$switchTime[switchDf$expId==ii]
   tipaPhaseRef(phaseRefTimes, switchTime)
})

phaseShifts = sapply(resultList, function(r) r$phaseShift)

write.csv(data.frame(expId = switchDf$expId, phaseShift = phaseShifts),
          'tipa_phaseref.csv', quote=FALSE, row.names=FALSE)
}
\seealso{
\code{\link{tipaCosinor}}
}
